Fixed the roll
Also made some small refactorings along the way.

The main problem was that the completely wrong number was being given to the RunnerObject move method (now renamed to SetVelocity).

Pretty much every number was calculated incorrectly so the code in the article must be wrong compared to the working demonstration gif they show.