Refactor states to not be scriptable objects
Using ScriptableObjects will work for a single player, but not more than one (hypothetically if we did that), but also it definitely won't work for enemies which there are multiple instances of each.

The reason why is complicated, but it has to do with the fact that have a copy of themselves for each instance, but are instead just a reference to a single location with the variables/code.

Instead they are just normal scripts that don't extend from anything (not even monobehavior). This is so that they don't clutter the inspector.