Changed logic of rollPressed variable
Previously rollPressed was set to true when receiving the roll event, and then the moving state would check this to transition to the roll.

The problem was that if you input this during idle, it would actually stay true and then the second you moved it would roll. Same thing if you double tap roll. So essentially the roll input was being buffered which is kinda not what we want.

Instead I changed the variable "rollPressed" to "rollInputDown" which follows the convention of Input.GetKeyDown() which is only true on the exact frame you started pressing it.

This requires a little extra work done in the update method but is  so much easier to use and fixes the above problems.