Fix sprite flipping to not occur during roll
The logic that made the sprite flip to face the correct direction was in the state machine, so it always ran no matter what state you were in.

This made the roll weird because you could change your sprite direction mid roll. This would also make attacking look weird later on as well.

I fixed it by instead calling the FlipSprite function inside the states that want to control it (in this case only PlayerMovingState)